pyproject.toml
[project]
name = "sms-weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "zai-sdk>=0.2.2",
]
